requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
WTForms==3.2.1
SQLAlchemy==2.0.36
python-dotenv==1.0.1
email-validator==2.2.0
